Removed unused "host" field. Split SettingsTestData.java to base JsonReader with children TestDataReader and EnvDataReader
