+++ src/test/java/tests/FileDownloadTest.java
@@ -1,16 +1,17 @@
 package tests;
 
 import pages.FileDownloadPage;
-import utils.SettingsTestData;
 import org.testng.annotations.AfterMethod;
 import org.testng.annotations.Test;
+import utils.JsonReader;
+import utils.TestDataReader;
 
 import java.io.File;
 
 public class FileDownloadTest extends BaseTest {
     private final FileDownloadPage fileDownloadPage = new FileDownloadPage();
-    private final String fileName = SettingsTestData.getFileData().getDownloadFile();
-    private final String filePath = SettingsTestData.RESOURCES_PATH + fileName;
+    private final String fileName = TestDataReader.getFileData().getDownloadFile();
+    private final String filePath = JsonReader.RESOURCES_PATH + fileName;
     private final File downloadedFile = new File(filePath);
 
     @Test
