+++ src/test/java/tests/FileUploadTest.java
@@ -1,11 +1,11 @@
 package tests;
 
-import utils.SettingsTestData;
 import org.testng.annotations.Test;
+import utils.TestDataReader;
 
 public class FileUploadTest extends BaseTest {
-    private final String fileName = SettingsTestData.getFileData().getUploadFile();
-    private final String filePath = SettingsTestData.TEST_DATA_PATH + fileName;
+    private final String fileName = TestDataReader.getFileData().getUploadFile();
+    private final String filePath = TestDataReader.TEST_DATA_PATH + fileName;
 
     @Test
     public void fileUploadTest() {
