+++ src/test/java/utils/FileUtils.java
@@ -12,7 +12,7 @@ public class FileUtils {
 
     public boolean isFileExist(File file) {
         try {
-            Awaitility.await().atMost(SettingsTestData.getEnvData().getWait(), TimeUnit.SECONDS).until(file::exists);
+            Awaitility.await().atMost(EnvDataReader.getEnvData().getWait(), TimeUnit.SECONDS).until(file::exists);
         } catch (ConditionTimeoutException exception) {
             return false;
         }
