+++ src/test/resources/environment/prod.json
@@ -1,6 +1,5 @@
 {
   "protocol": "https",
   "domain": "the-internet.herokuapp.com",
-  "wait": 10,
-  "host": "https://the-internet.herokuapp.com/"
+  "wait": 10
 }
